refactor: move gpu endpoints to modules
closes #1279
