+++ apps/api/src/gpu/repositories/day.repository.ts
@@ -1,0 +1,10 @@
+import { Day } from "@akashnetwork/database/dbSchemas/base";
+import { Op } from "sequelize";
+import { singleton } from "tsyringe";
+
+@singleton()
+export class DayRepository {
+  async getDaysAfter(date: Date) {
+    return await Day.findAll({ where: { date: { [Op.gte]: date } } });
+  }
+}
