+++ apps/api/src/gpu/routes/index.ts
@@ -1,0 +1,1 @@
+export * from "./gpu.router";
