+++ apps/api/src/gpu/types/gpu.type.ts
@@ -1,0 +1,51 @@
+import type { MsgCreateBid } from "@akashnetwork/akash-api/v1beta4";
+
+export type GpuProviderType = {
+  owner: string;
+  hostUri: string;
+  allocated: number;
+  allocatable: number;
+};
+
+export type GpuType = {
+  vendor: string;
+  model: string;
+  interface: string;
+  ram: string;
+  allocatable: number;
+  allocated: number;
+  providers: GpuProviderType[];
+  availableProviders: GpuProviderType[];
+};
+
+export type GpuBidType = {
+  height: number;
+  txHash: string;
+  datetime: Date;
+  provider: string;
+  aktTokenPrice: number;
+  hourlyPrice: number;
+  monthlyPrice: number;
+  deployment: {
+    owner: string;
+    cpuUnits: number;
+    memoryUnits: number;
+    storageUnits: number;
+    gpus: {
+      vendor: string;
+      model: string;
+      ram: string;
+      interface: string;
+    }[];
+  };
+  data: MsgCreateBid;
+};
+
+export type GpuWithPricesType = GpuType & {
+  prices: GpuBidType[];
+};
+
+export type ProviderWithBestBid = {
+  provider: GpuProviderType;
+  bestBid: GpuBidType;
+};
