+++ apps/api/src/routes/v1/index.ts
@@ -5,23 +5,6 @@ import proposals from "./proposals/list";
 import templateById from "./templates/byId";
 import templateList from "./templates/list";
 import templateListFull from "./templates/list-full";
-import gpu from "./gpu";
-import gpuBreakdown from "./gpuBreakdown";
-import gpuModels from "./gpuModels";
-import gpuPrices from "./gpuPrices";
 import trialProviders from "./trialProviders";
 
-export default [
-  address,
-  addressTransactions,
-  proposals,
-  proposalById,
-  templateListFull,
-  templateList,
-  templateById,
-  trialProviders,
-  gpu,
-  gpuModels,
-  gpuPrices,
-  gpuBreakdown
-];
+export default [address, addressTransactions, proposals, proposalById, templateListFull, templateList, templateById, trialProviders];
